+++ components/contracts/index.ts
@@ -1,2 +1,3 @@
 export { default as DocumentViewer } from './DocumentViewer';
 export { default as DocumentDiffViewer } from './DocumentDiffViewer';
+export { default as NewContractModal } from './NewContractModal';